TASK-4121 ENH: Also upload the Maven Site to Google Cloud
The site information isn't that interesting for MavenSetup itself, but projects that reuse the workflow should publish their site, as it contains the JavaDocs and dependency information.
If we'd delegate that to calling workflows, the site information would need to be cached or up/downloaded as artifacts because workspace files are not persisted across jobs.