Robustness: Deploy to Maven repo only after successful verification and site build
As deployment is part of the default lifecycle, it happens before the site (which is a different lifecycle) is built and staged.
A flaky connection to our Maven repo caused a build to fail while downloading Maven plugins required for the site build; i.e. the build failed _after_ the artifact had already been deployed.
Subsequent reruns then failed because the Maven artifact had already been deployed (and a Nexus repo doesn't allow a redeploy by default - the artifact needs to be manually deleted first):
> Error:  Failed to execute goal org.apache.maven.plugins:maven-deploy-plugin:3.1.2:deploy (default-deploy) on project juicyraspberrypie: Failed to deploy artifacts: Could not transfer artifact org.wensheng:juicyraspberrypie:pom:2.1.5 from/to reload (https://nexus.reloadkube.managedservices.resilient-teched.com/repository/reload/): status code: 400, reason phrase: Repository does not allow updating assets: reload (400) -> [Help 1]
It's better to first do all build steps (including the site), and only then do the deployment separately.