TASK-4335 BUG: Shaded plugin artifact is neither installed nor deployed
Both local and remote Maven repos just contain the unshaded JAR. The shaded JAR only exists locally in the target folder.
For JuicyRaspberryPie this wasn't important, as there are no external dependencies that need to be shaded so far. But we're going to consume the ReLoAd family plugins now via the Maven repo as well, and those do have shading.
To have both original and shaded artifact, it needs to be attached; cp. https://maven.apache.org/plugins/maven-shade-plugin/examples/attached-artifact.html
The configuration already enabled <shadedArtifactAttached>, but didn't specify the <shadedClassifierName>. (I couldn't find any documentation whether there's a default for that, but the example specifies it.)
Additionally, the (also not documented) <outputFile> prevents the install; cp. https://stackoverflow.com/a/67112650/813602